+++ CSV2Paper/main.py
@@ -1,0 +1,7 @@
+from tkinter import *
+from main_window import MainWindow
+
+if __name__ == '__main__':
+	base = Tk()
+	app = MainWindow(base)
+	base.mainloop()
