+++ CSV2Paper/resources/version.json
@@ -1,0 +1,1 @@
+{"exec_version":"0.9.2"}
